Finalizando o modal
